crawling/requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
schedule~=1.2.2
pytz~=2024.1
PyMySQL~=1.1.1
SQLAlchemy~=2.0.34